apiconvert/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
aniso8601==9.0.1
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==22.1.0
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloud-sql-python-connector==0.9.3
colorama==0.4.6
cryptography==38.0.3
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2
frozenlist==1.3.3
google-api-core==2.10.2
google-auth==2.14.1
google-cloud-core==2.3.2
google-cloud-pubsub==2.13.11
google-cloud-storage==2.6.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
greenlet==2.0.1
grpc-google-iam-v1==0.12.4
grpcio==1.50.0
grpcio-status==1.50.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-sqlalchemy==0.28.1
multidict==6.0.2
packaging==21.3
pg8000==1.29.3
proto-plus==1.22.1
protobuf==4.21.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
rsa==4.9
scramp==1.4.4
setuptools-scm==7.0.5
six==1.16.0
SQLAlchemy==1.4.44
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
yarl==1.8.1
